// src/interfaces/IStorageOptions.ts
export declare interface IStorageOptions {
  app: string
  version: string
  storage: string
  slide: string
  presentation?: string
  token?: string
  bearer?: string
  csrf?: string
  moderation?: string
  beforeTime?: string
  delay?: number
  period?: number
  retention?: number
}